+++ src/util/exception.cpp
@@ -17,5 +17,6 @@ namespace util
 {
 
 void exception::anchor() const {}
+void RuntimeError::anchor() const {}
 }
 }